test: split turn execution coverage by module
