+++ src/post-turn-pull-request.test.ts
@@ -1,0 +1,102 @@
+import test from "node:test";
+import assert from "node:assert/strict";
+import path from "node:path";
+import { handlePostTurnPullRequestTransitionsPhase } from "./post-turn-pull-request";
+import { PullRequestCheck, ReviewThread, SupervisorStateFile } from "./types";
+import { createConfig, createFailureContext, createIssue, createPullRequest, createRecord } from "./turn-execution-test-helpers";
+
+test("handlePostTurnPullRequestTransitionsPhase refreshes PR state after marking ready", async () => {
+  const config = createConfig();
+  const issue = createIssue({ title: "Refresh post-ready PR state" });
+  const draftPr = createPullRequest({ title: "Refresh after ready", isDraft: true });
+  const readyPr = createPullRequest({ title: "Refresh after ready" });
+  const initialChecks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
+  const postReadyChecks: PullRequestCheck[] = [{ name: "build", state: "IN_PROGRESS", bucket: "pending", workflow: "CI" }];
+  const state: SupervisorStateFile = {
+    activeIssueNumber: 102,
+    issues: { "102": createRecord({ last_head_sha: "head-115" }) },
+  };
+
+  let readyCalls = 0;
+  let snapshotLoads = 0;
+  let syncJournalCalls = 0;
+  const result = await handlePostTurnPullRequestTransitionsPhase({
+    config,
+    stateStore: {
+      touch: (record, patch) => ({ ...record, ...patch, updated_at: record.updated_at }),
+      save: async () => undefined,
+    },
+    github: {
+      getPullRequest: async () => {
+        throw new Error("unexpected getPullRequest call");
+      },
+      getChecks: async () => {
+        throw new Error("unexpected getChecks call");
+      },
+      getUnresolvedReviewThreads: async () => {
+        throw new Error("unexpected getUnresolvedReviewThreads call");
+      },
+      markPullRequestReady: async (prNumber: number) => {
+        assert.equal(prNumber, 116);
+        readyCalls += 1;
+      },
+    },
+    context: {
+      state,
+      record: state.issues["102"]!,
+      issue,
+      workspacePath: path.join("/tmp/workspaces", "issue-102"),
+      syncJournal: async () => {
+        syncJournalCalls += 1;
+      },
+      memoryArtifacts: {
+        alwaysReadFiles: [],
+        onDemandFiles: [],
+        contextIndexPath: "/tmp/context-index.md",
+        agentsPath: "/tmp/AGENTS.generated.md",
+      },
+      pr: draftPr,
+      options: { dryRun: false },
+    },
+    derivePullRequestLifecycleSnapshot: (_record, pr, checks) => ({
+      recordForState: _record,
+      nextState: checks.some((check) => check.bucket === "pending") ? "waiting_ci" : "pr_open",
+      failureContext: checks.some((check) => check.bucket === "pending")
+        ? null
+        : createFailureContext("unexpected failure"),
+      reviewWaitPatch: { review_wait_started_at: "2026-03-13T06:26:22Z", review_wait_head_sha: pr.headRefOid },
+      copilotRequestObservationPatch: {},
+      copilotTimeoutPatch: {
+        copilot_review_timed_out_at: null,
+        copilot_review_timeout_action: null,
+        copilot_review_timeout_reason: null,
+      },
+    }),
+    applyFailureSignature: () => ({
+      last_failure_signature: null,
+      repeated_failure_signature_count: 0,
+    }),
+    blockedReasonFromReviewState: () => null,
+    summarizeChecks: (checks) => ({
+      hasPending: checks.some((check) => check.bucket === "pending"),
+      hasFailing: checks.some((check) => check.bucket === "fail"),
+    }),
+    configuredBotReviewThreads: () => [],
+    manualReviewThreads: () => [],
+    mergeConflictDetected: () => false,
+    loadOpenPullRequestSnapshot: async () => {
+      snapshotLoads += 1;
+      return snapshotLoads === 1
+        ? { pr: draftPr, checks: initialChecks, reviewThreads: [] satisfies ReviewThread[] }
+        : { pr: readyPr, checks: postReadyChecks, reviewThreads: [] satisfies ReviewThread[] };
+    },
+  });
+
+  assert.equal(result.pr.isDraft, false);
+  assert.equal(result.record.state, "waiting_ci");
+  assert.equal(result.record.review_wait_head_sha, "head-116");
+  assert.equal(result.record.last_head_sha, "head-116");
+  assert.equal(readyCalls, 1);
+  assert.equal(snapshotLoads, 2);
+  assert.equal(syncJournalCalls, 0);
+});
